requirements.txt
pika==1.3.2
python-dotenv==1.0.0
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic>=2.0.0
pandas>=2.0.0
pyarrow>=12.0.0

